Clean up of where new alerts code is located
